% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uasparseplus.R
\name{uasparseplus}
\alias{uasparseplus}
\title{Parse User-Agent strings, providing extra information}
\usage{
uasparseplus(user_agents, ...)
}
\arguments{
\item{user_agents}{A vector of user-agent strings.}

\item{...}{Other parameters, passed to \code{uaparserjs::ua_parse()}.
\code{uaparserjs::ua_parse()} is documented to accept a \code{.progress}
parameter.}
}
\description{
\code{uasparseplus} parses User-Agent strings using
\code{uaparserjs::ua_parse()} to provide the bulk of its logic.
It augments the outputs of \code{uaparserjs::ua_parse()} with an additional
variable --- \code{device.type} --- which provides an indication of type of
device the User-Agent string represents.
}
\details{
The regex for identifying device types is derived almost entirely from the
code for Stata's \code{parseuas} package (version 1.4, 2020), with only
minor additions to cope with operating systems that are identified
differently by \code{uaparserjs::ua_parse()} and \code{parseuas}.
(Specifically, \code{uaparserjs::ua_parse()} identifies Ubuntu and Linux
as separate OSs, whereas \code{parseuas} classifies both as Linux.)

Because the core parsing functionality is provided by
\code{uaparserjs::ua_parse()}, which has a different set of regexes to
\code{parseuas}, the end results are not guaranteed to be identical
to those of \code{parseuas}.
}
\section{References}{


Roßmann, Joss, and Tobias Gummer. 2020.
  PARSEUAS: Stata Module to Extract Detailed Information from User Agent
  Strings (version 1.4). Chestnut Hill, MA: Boston College.

Roßmann, Joss, Tobias Gummer, and Lars Kaczmirek. 2020. ‘Working with User
  Agent Strings in Stata: The Parseuas Command’. Journal of Statistical
  Software 92 (1): 1–16. https://doi.org/10.18637/jss.v092.c01.

Rudis, Bob, Lindsey Simon, and Tobie Langel. 2020. Uaparserjs: Parse
  ‘User-Agent’ Strings (R package version 0.3.5).
  https://CRAN.R-project.org/package=uaparserjs.
}

\examples{
# Set up some sample data for parsing

uastrings <- c(
  paste0("Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/535.2 (KHTML, ",
         "like Gecko) Ubuntu/11.10 Chromium/15.0.874.106 ",
         "Chrome/15.0.874.106 Safari/535.2",
         collapse=""),

  paste0("Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 ",
         "(KHTML, like Gecko) Chrome/90.0.4430.212 Safari/537.36",
         collapse=""),

  paste0("Mozilla/5.0 (Linux; Android 9; moto g(6)) AppleWebKit/537.36 ",
         "(KHTML, like Gecko) Chrome/90.0.4430.210 Mobile Safari/537.36",
         collapse="")

)


# Parse the User-Agent strings

parsed_uastrings <- uasparseplus(uastrings)

# Examine the extra variable

table(parsed_uastrings$device.type)

}
